Reverted eval back to ints as its not neccessary for floats imo, FIXED A MAJOR BUG IN SEARCH FUNCTION +100 elo :)

+++ Evaluation/Eval.h
@@ -8,7 +8,7 @@
 using namespace chess;
 using namespace std;
 
-float Evaluation(const Board& board, int Ply);
+int Evaluation(const Board& board, int Ply);
 
 #endif
